sw64: cpufreq: refactor cpufreq related code
This refactoring centralizes the cpufreq drivers of Sunway platforms in
sunway-cpufreq.c, improving the maintainability of the code.
