--- arch/sw_64/platform/Makefile
@@ -1,2 +1,0 @@
-# SPDX-License-Identifier: GPL-2.0
-obj-y	+= cpufreq.o
